refactor: split dataService.ts into separate service modules
Extract LocalStorageService and SupabaseService classes into their own
files, reducing dataService.ts from 1,267 lines to a thin interface +
factory module (~45 lines). All public exports remain backward compatible.
